Looping issues when using new RESTful code
