--- src/main/java/controller/EditInventory.java
@@ -1,4 +1,0 @@
-package controller;
-
-public class EditInventory {
-}
