--- src/main/java/persistance/WorldsDao.java
@@ -1,4 +1,0 @@
-package persistance;
-
-public class WorldsDao {
-}
