+++ src/main/java/worldbuilder/controller/EditInventory.java
@@ -1,0 +1,4 @@
+package worldbuilder.controller;
+
+public class EditInventory {
+}
