+++ src/main/java/worldbuilder/entity/Inventories.java
@@ -1,4 +1,4 @@
-package entity;
+package worldbuilder.entity;
 
 /**
  * The type Inventories.
