+++ src/main/java/worldbuilder/entity/Owners.java
@@ -1,4 +1,4 @@
-package entity;
+package worldbuilder.entity;
 
 /**
  * The type Owners.
