+++ src/main/java/worldbuilder/persistance/Database.java
@@ -1,9 +1,8 @@
-package persistance;
+package worldbuilder.persistance;
 
-import utilities.PropertiesLoader;
+import worldbuilder.utilities.PropertiesLoader;
 
 import java.io.BufferedReader;
-import java.io.IOException;
 import java.io.InputStreamReader;
 import java.sql.Connection;
 import java.sql.DriverManager;
