+++ src/main/java/worldbuilder/persistance/InventoriesDao.java
@@ -1,6 +1,6 @@
-package persistance;
+package worldbuilder.persistance;
 
-import entity.Inventories;
+import worldbuilder.entity.Inventories;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.hibernate.Session;
@@ -9,7 +9,6 @@ import org.hibernate.Transaction;
 
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Expression;
 import javax.persistence.criteria.Root;
 import java.util.List;
