+++ src/main/java/worldbuilder/persistance/ItemsDao.java
@@ -1,10 +1,10 @@
-package persistance;
+package worldbuilder.persistance;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.hibernate.SessionFactory;
 
-public class OwnersDao {
+public class ItemsDao {
 
     private final Logger logger = LogManager.getLogger(this.getClass());
     SessionFactory sessionFactory = SessionFactoryProvider.getSessionFactory();
