+++ src/main/java/worldbuilder/persistance/LocationsDao.java
@@ -1,4 +1,4 @@
-package persistance;
+package worldbuilder.persistance;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
