+++ src/main/java/worldbuilder/persistance/UserDao.java
@@ -1,6 +1,6 @@
-package persistance;
+package worldbuilder.persistance;
 
-import entity.User;
+import worldbuilder.entity.User;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.hibernate.Session;
