+++ src/main/java/worldbuilder/persistance/WorldsDao.java
@@ -1,0 +1,4 @@
+package worldbuilder.persistance;
+
+public class WorldsDao {
+}
