+++ src/main/resources/log4j2.properties
@@ -1,5 +1,5 @@
 #property.filename = logs
-property.filename = C:/Users/frais/IdeaProjects/logs/FantasyWorldTracker
+property.filename = C:/Users/frais/IdeaProjects//FantasyWorldTracker/logs
 appenders = console, file
 
 appender.console.type = Console
@@ -14,7 +14,7 @@ appender.file.layout.type=PatternLayout
 appender.file.layout.pattern=[%-5level] %d{yyyy-MM-dd HH:mm:ss.SSS} [%t] %c{1} - %msg%n
 
 loggers=file
-logger.file.name=edu.matc
+logger.file.name=worldbuilder
 logger.file.level = debug
 logger.file.appenderRefs = file
 logger.file.appenderRef.file.ref = LOGFILE
